requirements.txt
rich
psycopg2-binary
python-dotenv
pymongo
neo4j
